+++ FabricTestExample/App.js
@@ -80,10 +80,10 @@ import Test1391 from './src/Test1391';
 import Test1419 from './src/Test1419';
 import Test1473 from './src/Test1473';
 import Test1476 from './src/Test1476';
-import TestSheet from './src/TestSheet';
+import Test1649 from './src/Test1649';
 
 enableFreeze(true);
 
 export default function App() {
-  return <TestSheet />;
+  return <Test42 />;
 }
